pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "general-mcp"
version = "0.1.0"
description = "Server MCP per la ricerca di sentenze della Corte di Cassazione italiana tramite ItalgiureWeb"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.1.1",
    "lxml>=6.1.0",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "readability-lxml>=0.8.4.1",
]
